Rename ifs tfs
