+++ src/library/math/vector_inline_functions.hpp
