+++ src/library/math/vector_template_functions.hpp
